Rename NormalizeHashDOTinit to NormalizeHashDOT_init
