+++ lib/NormalizeHashDOT_init.browser.js
@@ -637,4 +637,4 @@ sha256._init = function() {
 };
 
 })(); 
-export const NormalizeHashDOTinit = NormalizeHashDOTinit;
+export const NormalizeHashDOT_init = NormalizeHashDOT_init;
