+++ lib/NormalizeHashDOT_init.js
