+++ lib/jsonld.js
@@ -1,4 +1,4 @@
-import './NormalizeHashDOTinit.js';
+import './NormalizeHashDOT_init.js';
 import {jsonldDOTcompact} from './jsonldDOTcompact';
 export const compact = jsonldDOTcompact;
 import {jsonldDOTexpand} from './jsonldDOTexpand';
